Rewrote code base
I decided to rewrite the code base to be more optimized and (hopefully)
a little bit less confusing.

Simplified the "Player" class so it is not super long, it now only
changes player stats in the methods. The other functionality was moved
to new scripts suchs as "PlayerFascade"

Renamed WalkingScript to GPS. Also removed any references of Player
class in the GPS to make it easily portable to other projects and more
modular. GPS now depends on the "ObservedValue" class that was imported
using the "Gamelogic" extension. The deltaDistance variable is now an
ObservedValue object. See GameLogic extension for more info on how to
use it.

Added small scripts to deal with certain buttons clicks to simplify the
Manager classes.

Changed the attacking script to make it easier to read and made a new
function instead of copying and pasting code.

Also added in some comments and summaries of functions.