Removed all of the standard assets
Others:
Fixed a bug in the GPS script that would make it run on computers which
led to lots of errors.
Fixed a bug in persisentUIElements that would lead to null pointer
errors in some scenarios
Adjusted build settings for web build so that you'd be able to see the
entire screen.