ARMOR! plus enemy item drops
Made armor almost work except for A REALLY SUPID AND DUMB AND IDIOTIC
AND MORINIC NULLREFFERENCEEXCEPTION!!!! But theres a bunch of armor
sprites now, and enemies can drop items now, except the sprites arent
done yet so they teeny tiny. Also made it so the game pauses when you
open up the journal.