Many small fixes
Enemies are added to the queue based on distance walked value changes
rather than update.
Overlapping items no longer matter.
Remove the "Distance" attribute of player because it was never used and
I have no idea what it was for.
Enemies can be added to the enemy queue after reaching the end of the
quest.
Added a boss fight tutorial.
Queue number next to enemy head on walking screen now updates whenever
the player moves, instead of every 2 seconds.